admin role now required for protected endpoints
The admin role is now neccessary to create, delete and update
categories, and admins can now delete listings that are not their own

# Conflicts:
#	src/main/kotlin/org/ntnu/grepapp/controller/ListingController.kt

+++ src/main/kotlin/org/ntnu/grepapp/controller/CategoryController.kt
@@ -5,6 +5,9 @@ import org.ntnu.grepapp.model.Category
 import org.ntnu.grepapp.service.CategoryService
 import org.apache.logging.log4j.LogManager
 import org.ntnu.grepapp.dto.CategoryDTO
+import org.ntnu.grepapp.security.JwtUtil
+import org.ntnu.grepapp.security.JwtUtil.*
+import org.ntnu.grepapp.service.AuthService
 import org.springframework.data.domain.PageRequest
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -15,6 +18,8 @@ import org.springframework.web.bind.annotation.*
 @CrossOrigin(origins = ["http://localhost:5173"])
 class CategoryController(
     private val categoryService: CategoryService,
+    private val authService: AuthService,
+    private val jwtUtil: JwtUtil,
 ) {
     private val logger = LogManager.getLogger(this::class::java);
 
@@ -29,6 +34,9 @@ class CategoryController(
 
     @PostMapping("/create")
     fun create(@RequestBody request: CategoryCreateRequest): ResponseEntity<Unit> {
+        val role = authService.getRole()
+        if (role != "admin") { return ResponseEntity(HttpStatus.FORBIDDEN) }
+
         val category = Category(request.name);
         val status = if (categoryService.create(category)) {
             HttpStatus.OK
@@ -40,6 +48,9 @@ class CategoryController(
 
     @PatchMapping("/update")
     fun update(@RequestBody request: CategoryUpdateRequest): ResponseEntity<Unit> {
+        val role = authService.getRole()
+        if (role != "admin") { return ResponseEntity(HttpStatus.FORBIDDEN) }
+
         val new = Category(request.new.name);
         val status = if (categoryService.update(request.oldName, new)) {
             HttpStatus.OK
@@ -51,6 +62,9 @@ class CategoryController(
 
     @DeleteMapping("/delete/{name}")
     fun delete(@PathVariable name: String): ResponseEntity<Unit> {
+        val role = authService.getRole()
+        if (role != "admin") { return ResponseEntity(HttpStatus.FORBIDDEN) }
+
         val status = if (categoryService.delete(name)) {
             HttpStatus.OK
         } else {
